Using 2.1's configuration.
Anki 2.1 use a new way to configure add-on.
This new configuration allow to easily transfer the configuration from
a computer to another one. Which is exactly what I wanted to do.

This also add the possibility to choose that the number of column
is the same in every window (by default, it remains distinct in
browser, advanced browser, add window, etc...)